client/requirements.txt
mcp-use
langchain-openai
python-dotenv